Move CDiskTxPos/CBlockUndo to txdb.h/undo.h respectively
